+++ web-experimental/src/resources/product/components/ProductLayout.jsx
@@ -2,7 +2,6 @@
  * Wraps all Product views in a wrapping container. If you want to give all
  * product views a sidebar for example, you would set that here.
  * 
- * TODO:
  * Accepts a "title" prop and passes it down to be used by React Helmet
  * This allows us to easily update the browser tab title on each view.
  */
