+++ web-experimental/src/resources/product/views/UpdateProduct.jsx
@@ -1,54 +1,65 @@
 // import primary libraries
 import React from 'react'
+// import PropTypes from 'prop-types'; // this component gets no props
 import { useParams, useHistory } from 'react-router-dom'
 
-// import the hooks generated by RTK query
-import { useSingleProduct, useUpdateProduct } from '../productService';
-
-import AsyncWrapper from '../../../global/components/helpers/AsyncWrapper';
-
+// import global components
+import WaitOn from '../../../global/components/helpers/WaitOn';
 
 // import resource components
 import ProductLayout from '../components/ProductLayout.jsx'
 import ProductForm from '../components/ProductForm.jsx';
 
-const UpdateProduct = () => {
+// import services
+import { useGetUpdatableProduct } from '../productService';
 
+const UpdateProduct = () => {
   const history = useHistory(); // get history object
   const { productId } = useParams(); // replaces match.params.productId
 
-  // use the hook to fetch the single product
-  const { data: product, ...productFetch } = useSingleProduct(productId);
+  // fetches and returns the product and the update action wrapped in dispatch.
+  // another benefit of using this version is that productQuery.isFetching will be true while the update is being processed by the server.
+  const { sendUpdateProduct, data: product, ...productQuery } = useGetUpdatableProduct(productId);
 
-  // access the update action by running the mutation hook created in productService
-  // It returns an array where the first item is the update action and the second is an object with information about the result of the update action
-  const [
-    sendUpdateProduct, // this is the function that fires the POST call to api/products/:productId. Think of it as importing productActions.sendUpdateProduct
-    { isLoading: isUpdating }, // This is the destructured update result. Rename isLoading to isUpdating for clarity.
-  ] = useUpdateProduct();
-
-  const handleFormSubmit = async (updatedProduct) => {
+  const handleFormSubmit = updatedProduct => {
     // send the updatedProduct to the server
-    const { data: product } = await sendUpdateProduct(updatedProduct); // replaces dispatch(productActions.sendUpdateProduct(updatedProduct)).then(productRes => {...})
-    // back to single product view
-    history.push(`/products/${product._id}`)
+    sendUpdateProduct(updatedProduct);
+    // back to single product view. We don't have to wait for the update to finish. It's okay if the product is still updating when the user gets to the single product view.
+    history.push(`/products2/${updatedProduct._id}`);
   }
-  
+
   return (
+    // <ProductLayout title={'Update Product'}>
+    //   { productQuery.isLoading ? <div>Loading...</div>
+    //     : productQuery.isError ? <div>An error occurred 😬 <Button onClick={productQuery.refetch}>Refetch</Button></div>
+    //     : productQuery.isEmpty ? <div>Empty</div>
+    //     : // we have the product
+    //     <ProductForm
+    //       product={product}
+    //       cancelLink={`/products2/${productId}`}
+    //       disabled={productQuery.isFetching}
+    //       formTitle="Update Product"
+    //       formType="update"
+    //       handleFormSubmit={handleFormSubmit}
+    //     />
+    //   }
+    // </ProductLayout>
+
+    // <WaitOn/> handles all of the isLoading, isError, etc stuff so we don't have to do the stuff above
     <ProductLayout title={'Update Product'}>
-      <AsyncWrapper {...productFetch}>
-        { !product ? <div>No product found</div>
-          : // we have the product, render the form
+      <WaitOn query={productQuery}>
+        { product &&
+          // we have the product, render the form
           <ProductForm
             product={product}
-            cancelLink="/products"
-            disabled={productFetch.isFetching || isUpdating}
-            formTitle="Update Product"
+            cancelLink={`/products2/${productId}`}
+            disabled={productQuery.isFetching}
+            // formTitle="Update Product"
             formType="update"
             handleFormSubmit={handleFormSubmit}
           />
         }
-      </AsyncWrapper>
+      </WaitOn>
     </ProductLayout>
   )
 }
